requirements.txt
torch>=2.0.0
transformers>=4.30.0
biopython>=1.80
scikit-learn>=1.2.0
pandas>=2.0.0
requests>=2.28.0
matplotlib>=3.5.0
seaborn>=0.12.0
tqdm>=4.60.0
